requirements.txt
torch
torchvision
Pillow
timm
